Directory works now
The Directory command works now, in a similar fashion to the Help
command. The code I used for the directory seems flexible, as it can do
any directory signified by a string parameter. It may have a
larger-than-required runtime if I go to print the entire computer's
directory, but it seems do-able.